+++ NEWS.md
@@ -1,3 +1,9 @@
+# secretbase 0.5.0
+
+* Adds Keccak cryptographic hash algorithm.
+* Adds `shake256()` to delineate from `sha3()`.
+* Use of `sha3()` supplying 'bit' argument other than 224, 256, 384 or 512 is deprecated.
+
 # secretbase 0.4.0
 
 * Adds HMAC generation to `sha256()`.